services/model-proxy/requirements.txt
litellm==1.40.0
grpcio
grpcio-tools